+++ Runtime/Scripts/Schema/TextureExtension.cs
@@ -17,9 +17,6 @@ namespace GLTFast.Schema {
 
     [System.Serializable]
     public class TextureExtension {
-
-        [System.Obsolete("to be replaced by KHR_texture_basisu")]
-        public TextureFormat KHR_texture_cttf = null;
         public TextureFormat KHR_texture_basisu = null;
     }