+++ ChangeLog.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Correct shading in linear color space projects due to correct (linear) sampling of normal, occlusion and metallic-roughness maps
 - Memory leak: free up volatile array `imageFormats`
+### Changed
+- Removed support for obsolete draft extensions `KHR_texture_cttf` and `KHR_image_ktx2`
 
 ## [1.0.0] - 2020-03-13
 ### Changed